requirements.txt
albumentations==0.4.5
attrdict==2.0.1
dropbox==10.2.0
fire==0.3.1
jupyterlab==2.1.4
matplotlib==3.2.1
numpy==1.18.5
opencv-python==4.2.0.34
pandas==1.0.4
Pillow==6.2.2
pytorch-toolbelt==0.3.2
scikit-learn==0.23.1
seaborn==0.10.1
segmentation-models-pytorch==0.1.0
tensorboard==2.2.2
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.1
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
